Add metadata handling and content type determination for R2 uploads
